uint_8t is enough

+++ src/libvgcode/src/OptionTemplate.cpp
@@ -21,7 +21,7 @@ void OptionTemplate::init(uint8_t resolution)
         return;
 
     m_resolution = std::max<uint8_t>(resolution, 3);
-    m_vertices_count = static_cast<uint16_t>(3 * m_resolution);
+    m_vertices_count = static_cast<uint8_t>(3 * m_resolution);
     const float step = 2.0f * PI / float(m_resolution);
 
     //
